+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 This is a history of changes to clara-rules.
 
+# 0.23.0-SNAPSHOT
+* extract clara.rules.compiler/compile-test-handler from clara.rules.compiler/compile-test
+* add support for `env` inside of test expressions
+
 # 0.22.1
 * fix incorrent lint warning triggered when this binding is not used in clj-kondo hooks